refactor: clean project structure
