+++ src/client/generic.rs
