+++ src/lib.rs
@@ -1,3 +1,2 @@
 pub mod client;
 pub mod model;
-pub mod utils;
